+++ nomeroff_net/pipelines/number_plate_text_reading.py
@@ -19,15 +19,15 @@ DEFAULT_PRESETS = {
         "for_regions": ["eu", "xx_transit", "xx_unknown"],
         "model_path": "latest"
     },
-    "ru_shufflenet_v2_x2_0": {
+    "ru": {
         "for_regions": ["ru", "eu_ua_ordlo_lpr", "eu_ua_ordlo_dpr"],
         "model_path": "latest"
     },
-    "kz_shufflenet_v2_x2_0": {
+    "kz": {
         "for_regions": ["kz"],
         "model_path": "latest"
     },
-    "kg_shufflenet_v2_x2_0": {
+    "kg": {  # "kg_shufflenet_v2_x2_0"
         "for_regions": ["kg"],
         "model_path": "latest"
     },
@@ -39,7 +39,7 @@ DEFAULT_PRESETS = {
         "for_regions": ["su"],
         "model_path": "latest"
     },
-    "am_shufflenet_v2_x2_0": {
+    "am": {
         "for_regions": ["am"],
         "model_path": "latest"
     },
@@ -62,11 +62,13 @@ class NumberPlateTextReading(Pipeline):
                  default_label: str = "eu_ua_2015",
                  default_lines_count: int = 1,
                  class_detector=TextDetector,
+                 need_preprocess=False,
                  **kwargs):
         if presets is None:
             presets = DEFAULT_PRESETS
         super().__init__(task, image_loader, **kwargs)
         self.detector = class_detector(presets, default_label, default_lines_count)
+        self.need_preprocess = need_preprocess
 
     def sanitize_parameters(self, **kwargs):
         return {}, {}, {}
@@ -75,14 +77,17 @@ class NumberPlateTextReading(Pipeline):
         return super().__call__(images, **kwargs)
 
     def preprocess(self, inputs: Any, **preprocess_parameters: Dict) -> Any:
-        images, labels, lines = unzip(inputs)
+        images, labels, lines, preprocessed_np = unzip(inputs)
         images = [self.image_loader.load(item) for item in images]
-        return unzip([images, labels, lines])
+        return unzip([images, labels, lines, preprocessed_np])
 
     @no_grad()
     def forward(self, inputs: Any, **forward_parameters: Dict) -> Any:
-        images, labels, lines = unzip(inputs)
-        model_inputs = self.detector.preprocess(images, labels, lines)
+        images, labels, lines, preprocessed_np = unzip(inputs)
+        if self.need_preprocess or all([np is None for np in preprocessed_np]):
+            model_inputs = self.detector.preprocess(images, labels, lines)
+        else:
+            model_inputs = self.detector.preprocess(preprocessed_np, labels, lines, need_preprocess=False)
         model_outputs = self.detector.forward(model_inputs)
         model_outputs = self.detector.postprocess(model_outputs)
         return unzip([images, model_outputs, labels])