+++ nomeroff_net/pipes/number_plate_classificators/options_detector.py
@@ -75,8 +75,8 @@ class OptionsDetector(object):
             options = dict()
 
         # input
-        self.height = 64
-        self.width = 295
+        self.height = 50
+        self.width = 200
         self.color_channels = 3
 
         # outputs 1
@@ -398,9 +398,8 @@ class OptionsDetector(object):
             confidences.append([region_confidence, count_lines_confidence])
         return confidences, region_ids, count_lines
 
-    @staticmethod
-    def preprocess(images):
-        x = [normalize_img(img) for img in images]
+    def preprocess(self, images):
+        x = [normalize_img(img, height=self.height, width=self.width) for img in images]
         x = np.moveaxis(np.array(x), 3, 1)
         return x
 
@@ -415,7 +414,7 @@ class OptionsDetector(object):
         """
         Predict options(region, count lines) with confidence by numberplate images
         """
-        xs = [normalize_img(img) for img in imgs]
+        xs = [normalize_img(img, height=self.height, width=self.width) for img in imgs]
         if not bool(xs):
             return [], [], [], []
         predicted = self._predict(xs)